lib: move telegram-continuation-token to agent/lib
Canonical home moves from scripts/lib to agent/lib so the eve bundle
no longer reaches back into scripts/. Content is unchanged; only
importers' specifiers are updated.
